requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
async-websocket-client==0.1.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
frozenlist==1.4.0
gevent==23.9.0.post1
greenlet==2.0.2
idna==3.4
multidict==6.0.4
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
websocket==0.2.1
websockets==11.0.3
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
